+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 *
 
 ### Changed
-*
+* Fix at_hash verification #200
 
 ### Removed
 *