+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 * Fix at_hash verification #200
 * Getters for public parameters #204
+* Adding Token Request Parameters
 
 ### Removed
 * Removed explicit content-length header - caused issues with proxy servers